Add JSONSchemas and reorganize a bit of code
