+++ cmd/main.go
@@ -1,13 +1,10 @@
 package main
 
 import (
-	"encoding/json"
 	"os"
 
 	"github.com/alecthomas/kong"
 	"github.com/charmbracelet/log"
-	"github.com/pkg/errors"
-	"github.com/sayden/counters"
 )
 
 var logger = log.New(os.Stderr)
@@ -23,6 +20,8 @@ var Cli struct {
 	Vassal vassal `cmd:"" help:"Create a vassal module for testing. It searches for the 'template.xml' in the same folder"` //FIXME
 
 	GenerateTemplate GenerateTemplate `cmd:"" help:"Generates a new counter template file with default values"`
+
+	CheckTemplate CheckTemplate `cmd:"" help:"Check if a JSON file is a valid counter template"`
 }
 
 func main() {
@@ -35,109 +34,5 @@ func main() {
 	err := ctx.Run()
 	ctx.FatalIfErrorf(err)
 
-	log.Info("Done")
-}
-
-type GenerateTemplate struct {
-	OutputPath string `help:"Path to the folder to write the JSON" short:"o"`
-}
-
-func (i *GenerateTemplate) Run(ctx *kong.Context) error {
-	if i.OutputPath == "" {
-		return errors.New("output path is required")
-	}
-
-	err := generateNewCounterTemplate(i.OutputPath)
-	if err != nil {
-		return errors.Wrap(err, "could not generate new counter template")
-	}
-
-	return nil
-}
-
-func generateNewCounterTemplate(outputPath string) error {
-	counterTemplate := counters.CounterTemplate{
-		Counters: []counters.Counter{
-			{
-				Texts: []counters.Text{
-					{
-						Settings: counters.Settings{
-							Position: 3,
-						},
-						String: "String",
-					},
-				},
-				Images: []counters.Image{
-					{
-						Settings: counters.Settings{
-							Position: 0,
-						},
-						Path: "path/to/image",
-					},
-				},
-			},
-		},
-		Rows:                      7,
-		Columns:                   7,
-		Mode:                      "template",
-		OutputFolder:              "generated",
-		PositionNumberForFilename: 0,
-		Prototypes: map[string]counters.CounterPrototype{
-			"prototype1": {
-				TextsPrototypes: []counters.TextPrototype{
-					{
-						StringList: []string{"String1", "String2"},
-					},
-				},
-				ImagesPrototypes: []counters.ImagePrototype{
-					{
-						Image: counters.Image{
-							Settings: counters.Settings{
-								Position: 0,
-							},
-						},
-						PathList: []string{"path/to/image1", "path/to/image2"},
-					},
-				},
-				Counter: counters.Counter{
-					Texts: []counters.Text{
-						{
-							Settings: counters.Settings{
-								Position: 3,
-							},
-							String: "String",
-						},
-					},
-					Images: []counters.Image{
-						{
-							Settings: counters.Settings{
-								Position: 0,
-							},
-							Path: "path/to/image",
-						},
-					},
-				},
-			},
-		},
-		Settings: counters.Settings{
-			FontPath:        "BebasNeue-Regular.ttf",
-			Width:           100,
-			Height:          100,
-			Margins:         3,
-			FontHeight:      10,
-			FontColorS:      "black",
-			BackgroundColor: "white",
-			ImageScaling:    "fitWidth",
-		},
-	}
-
-	f, err := os.Create(outputPath)
-	if err != nil {
-		return errors.Wrap(err, "could not create output file")
-	}
-	defer f.Close()
-
-	enc := json.NewEncoder(f)
-	enc.SetIndent("", "  ")
-	return enc.Encode(counterTemplate)
+	logger.Info("Done")
 }
