--- schema.json
@@ -1,50 +1,0 @@
-{
-    "$schema": "https://json-schema.org/draft/2020-12/schema",
-    "$id": "https://example.com/product.schema.json",
-    "properties": {
-        "background_color": {
-            "type": "string",
-            "description": "The background color for the entire card. You can use a hexadecimal format or a string format like 'red', 'blue'"
-        },
-        "border_color": {
-            "type": "string",
-            "description": "The border color for the card. You can use a hexadecimal format or a string format like 'red', 'blue'"
-        },
-        "border_width": {
-            "type": "integer",
-            "description": "The width of the border in the card. 0 value means no border"
-        },
-        "cards": {
-            "type": "array",
-            "items": {
-                "type": "object",
-                "properties": {
-                    "multiplier": {
-                        "type": "integer",
-                        "description": "How many copies to generate of this specific entity"
-                    },
-                    "texts": {
-                        "type": "array",
-                        "items": {
-                            "type": "object",
-                            "properties": {
-                                "string": {
-                                    "type": "string",
-                                    "description": "a desc"
-                                },
-                                "position": {
-                                    "type": "integer",
-                                    "description": "A number from 0 to 16. The position of the text inside the container"
-                                },
-                                "shadow": {
-                                    "type": "integer",
-                                    "description": "If not 0, the size of the shadow for this text"
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
